move phone edit UI to ProfilePage, show read-only with profile link in EnrolledPanel
